producer/requirements.txt
stomp.py
confluent-kafka